Create GCP integration
An example of how you could use Google Cloud Platform (GCP) to analyze the contents of a user-generated image and determine the material composition of the object in the image using Python.

In this example, we use the google.cloud.vision library to create a Vision client and use it to detect and analyze the objects in the user-generated image. The object_localization method returns a list of LocalizedObjectAnnotation objects, which contain information about the objects detected in the image, including their material composition. We can iterate through the list of objects and print the name attribute to get the material composition of the object in the image.