Create TensorFlow & PyTorch material classification
This code uses TensorFlow and PyTorch to classify the material of the object in the user-generated image. It first imports the necessary libraries, including TensorFlow and PyTorch. The code then loads the user-generated image into a variable and uses TensorFlow and PyTorch to run the image through pre-trained machine learning models that have been trained to classify the materials of objects. The code then returns the material classification to CleanApp, which can use this information to provide information about the object in the image to the user.