Create Tesla navi popup
This code uses the matplotlib and seaborn libraries to create a heatmap of the CleanApp reports, and then displays the heatmap in a popup notification on the Tesla navigation display. It also includes a check to ensure that there are more than 100 reports in the past 1 hour before displaying the notification.

Note that this is just sample code and may need to be modified to fit the specific requirements of your use case. For example, you may need to modify the is_within_500m function to accurately calculate the distance between two locations, and you may need to add additional code to notify first responders.